RHEL/CentOS: update VOLUME to match Documentation instructions
The Documentation for mssql in Docker uses the /var/opt/mssql mount point
while the Dockerfile is using the /var/opt/mssql/data mount point.

If the user specifies a persistent volume for the documented mount point,
the data may not be persisted due to Docker mounting the data dir separately
to non-persistent storage.